Add method that handles binary payloads to MessageFormatUtils
